♻️ Simplify i18n: remove namespace param from useI18n, migrate www to Astro i18n
