+++ apps/desktop/src/components/UpgradeProgress.vue
@@ -7,7 +7,7 @@ defineProps<{
   progress: number
 }>()
 
-const { t } = useI18n('desktop')
+const { t } = useI18n()
 const model = defineModel<boolean>({ required: true })
 </script>
