+++ apps/desktop/src/composables/useI18n.ts
@@ -2,7 +2,6 @@
 import type { Locale } from '@typo/languages'
 import { emit, listen } from '@tauri-apps/api/event'
 import { createGenericTranslator, defaultLocale } from '@typo/languages'
-import { messages as sharedMessages } from '@typo/languages/messages/common'
 import { computed, ref } from 'vue'
 import { get, save, set } from '@/store'
 import en from '../locales/en.json'
@@ -29,16 +28,8 @@ export async function setLocale(next: Locale): Promise<void> {
   await emit(LOCALE_EVENT, next)
 }
 
-const allMessagesForLocale = {
-  en: { ...sharedMessages.en, ...localMessages.en },
-  zh: { ...sharedMessages.zh, ...localMessages.zh },
-  jp: { ...sharedMessages.jp, ...localMessages.jp },
-}
-
-export function useI18n(_namespace?: string) {
-  // We ignore namespace for now as we merge all local keys into one bundle per app
-  // This simplifies the structure as requested
-  const translator = computed(() => createGenericTranslator(locale.value, allMessagesForLocale))
+export function useI18n() {
+  const translator = computed(() => createGenericTranslator(locale.value, localMessages))
 
   const t = (key: string, vars?: Record<string, string | number | undefined | null>) => translator.value(key, vars)
