+++ apps/desktop/src/windows/Upgrade.vue
@@ -7,7 +7,7 @@ import { useGlobalState } from '@/composables/useGlobalState'
 import { useI18n } from '@/composables/useI18n'
 
 const { updateInfo, setCurrentWindow } = useGlobalState()
-const { t } = useI18n('desktop')
+const { t } = useI18n()
 
 const isUpgrading = ref(false)
 const downloadProgress = ref(0)
