+++ apps/www/src/components/FeatureGrid.astro
@@ -1,7 +1,5 @@
 ---
-import { tr } from '@/lib/i18n'
-
-const t = tr(Astro, 'www')
+const { t } = Astro.locals
 
 const features = [
   {
