+++ apps/www/src/components/Footer.astro
@@ -1,8 +1,8 @@
 ---
-import { getLocale, getLocalizedPath, tr } from '@/lib/i18n'
+import { getLocale, getLocalizedPath } from '@/lib/i18n'
 
-const t = tr(Astro, 'www')
-const currentLocale = getLocale(Astro)
+const { t } = Astro.locals
+const currentLocale = getLocale(Astro.currentLocale)
 const year = new Date().getFullYear()
 const githubLink = `<a href="https://github.com/yuler/typo" class="font-medium text-zinc-800 underline-offset-2 hover:underline dark:text-zinc-200">${t('footer.github_label')}</a>`
 const copyrightHTML = t('footer.copyright', { year, github: githubLink })
