+++ apps/www/src/components/Hero.astro
@@ -1,8 +1,8 @@
 ---
-import { tr, getLocalizedPath, getLocale } from '@/lib/i18n'
+import { getLocalizedPath, getLocale } from '@/lib/i18n'
 
-const t = tr(Astro, 'www')
-const locale = getLocale(Astro)
+const { t } = Astro.locals
+const locale = getLocale(Astro.currentLocale)
 ---
 
 <section class="mx-auto max-w-5xl px-4 pb-16 pt-12 sm:px-6 sm:pt-16">
