+++ apps/www/src/components/Testimonials.astro
@@ -1,8 +1,8 @@
 ---
-import { tr, getLocale, getLocalizedPath } from '@/lib/i18n'
+import { getLocale, getLocalizedPath } from '@/lib/i18n'
 
-const t = tr(Astro, 'www')
-const locale = getLocale(Astro)
+const { t } = Astro.locals
+const locale = getLocale(Astro.currentLocale)
 
 type Testimonial = {
   quote: string
