+++ apps/www/src/pages/index.astro
@@ -5,9 +5,7 @@ import Header from '@/components/Header.astro'
 import Hero from '@/components/Hero.astro'
 import Testimonials from '@/components/Testimonials.astro'
 import BaseLayout from '@/layouts/BaseLayout.astro'
-import { tr } from '@/lib/i18n'
-
-const t = tr(Astro, 'www')
+const { t } = Astro.locals
 ---
 
 <BaseLayout title={t('page.home.title')}>
